refactor: split cli modules and tests
