+++ src/lifeos_cli/cli_support/__init__.py
@@ -1,0 +1,2 @@
+"""Internal CLI support modules for lifeos_cli."""
+
